requirements.txt
Flask
requests
beautifulsoup4
google-cloud-bigquery
gunicorn